backend/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
python-socketio>=5.10.0
python-engineio>=4.8.0
eventlet>=0.33.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.26.0
scipy>=1.11.0
PyJWT>=2.8.0
websockets>=12.0
